+++ uui-build/ts/figmaTokensGen/config.ts
@@ -1,9 +1,0 @@
-import { ITokensConfig } from './types/configTypes';
-
-export const CONFIG: ITokensConfig = {
-    tokens: {
-        '--uui-text-contrast': { supported: false },
-        '--uui-text-white': { supported: false },
-        '--uui-text-black': { supported: false },
-    },
-};