+++ uui-core/src/types/components/datePicker/BaseRangeDatePicker.ts
@@ -27,9 +27,9 @@ export type RangeDatePickerPresets = {
 
 export type RangeDatePickerPresetValue = {
     /** Range from value */
-    from: string;
+    from?: string;
     /** Range to value */
-    to: string;
+    to?: string;
     /** Preset order in presets list */
     order?: number;
 };