+++ uui-editor/src/plugins/superscriptPlugin/superscriptPlugin.tsx
@@ -5,14 +5,15 @@ import { isPluginActive } from '../../helpers';
 import { ReactComponent as SuperScriptIcon } from '../../icons/super-script.svg';
 import { ToolbarButton } from '../../implementation/ToolbarButton';
 import { handleMarkButtonClick } from '../../utils/handleMarkButtonClick';
-import { IHasToolbarButton } from "../../implementation/Toolbars";
+import { IHasToolbarButton } from '../../implementation/Toolbars';
 
 const SUPERSCRIPT_TYPE = 'uui-richTextEditor-superscript';
 
 export const superscriptPlugin = () => createSuperscriptPlugin<IHasToolbarButton>({
     type: SUPERSCRIPT_TYPE,
     options: {
-        floatingBarButton: SuperscriptButton
+        floatingBarButton: SuperscriptButton,
+        name: 'superscription-button',
     },
 });